pyproject.toml
[tool.poetry]
name = "landolfio"
version = "0.1.0"
description = "Landolfio Asset Tracking"
authors = ["Job Doesburg <job.doesburg@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.5"
Pillow = "^7.1.2"
django-admin-autocomplete-filter = "^0.5"
django-import-export = "^2.0.2"
requests = "^2.23.0"
uwsgi = "^2.0.18"
moneybird = "^0.1.3"
django-countries = "^6.1.2"
django-iban-field = "^0.8"
django-polymorphic = "^2.1.2"

[tool.poetry.dev-dependencies]
pydocstyle = "^5.0.2"
black = "^19.10b0"
flake8 = "^3.7.9"

[tool.black]
line-length = 119
target-version = ["py37"]
exclude = '''
/(
  migrations
  | website/landolfio/settings
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
